requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-extensions==3.2.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mypy==1.2.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.19.1
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
yapf==0.32.0
